refactor(vacs-client): extract navigation logic into router component
